updated readme
Attendance is a repository containing code involved in the research of zero-effort attendance tracking solutions. Current methodology involves usage of in-place Wi-Fi access points and their transmitted ambient signals to locate devices inside of buildings. The proposed system aims to have students¿ smartphones in the classroom to generate a ¿location proof" based on the radio frequency fingerprints scanned by the user¿s device that could then be used to verify users' locations. Data collected from user and networking devices are hashed for user privacy before reaching the application server for users' privacy. Thus, the server has no information about the identity of the students nor where the class is taking place.

The proposed system consists of users' devices and the application server. Then, we have the  Wi-Fi access points in the infrastructure. The users¿ devices scan Wi-Fi access points in the area. Then, the MAC addresses of the Wi-Fi access points are hashed then uploaded along with the Received Signal Strength Indicators (RSSI) to the application server.

The users¿ devices include the students and the instructor smartphones with the mobile application installed. On the mobile app, there are two types of users: the students and the instructor. Each student user has a unique userID, which is generated by hashing the student name with a salt, which is the course number. During the class, the mobile app performs a Wi-Fi scan which returns nearby access points MAC addresses, and their measured RSSIs.

The application server calculates the Euclidian distance between all student data and the instructor datum. If the distance is below a certain defined threshold, the attendance decision is present. If the distance is above the threshold, the attendance decision is absent.

<p align="center">
    <img width="400" src="https://raw.githubusercontent.com/awshene/Attendance/main/system_architecture.png" alt="Proposed System Architecture"/>
</p>

The experiment took place inside a classroom in a building on campus. Data was collected in a total of 15 different locations in the building. Nine from inside the classroom and six from other locations in the building as shown in the figure below. Data collection was conducted using an android device with an application that scans for Wi-Fi access points and uploads the result to a server. A data point was collected every minute for one hour to provide sufficient data to analyze. To determine whether the student device is in the attendance range, the instructor device was set as a reference point.

<p align="center">
    <img width="600" src="https://raw.githubusercontent.com/awshene/Attendance/main/floormap.png" alt="Floor Map"/>
</p>

After data collection, the ten access points with the highest average RSSIs were selected from the instructor podium as the reference set. Corresponding averages for the access points from the other datasets were aggregated into a single table to train the classifier. Visualization of the recorded data is shown in the figure below. Then, the Euclidean distance was found for each data set and reference point. The Euclidian distance between each test location and the reference set was calculated as well.

<p align="center">
    <img width="500" src="https://raw.githubusercontent.com/awshene/Attendance/main/apdata.png" alt="Access Point Data"/>
</p>

To test our classifier, 198 test data points were collected. The Euclidean distance was calculated between each of the data points and the reference set. A threshold of 56.5 is considered to classify if the data point is inside or outside the classroom. The test results are shown in the figure below. Error reporting on the classification demonstrates that predictions for locations truly inside of the classroom were correct 100 percent of the time, while predictions for locations truly outside of the classroom were correct 98.75 percent of the time. The accounted for error is a result of the measurements recorded from the location at the door, which triggered a false positive in the classification due to its  proximity to the classroom itself and the instructor podium.

<p align="center">
    <img width="400" src="https://raw.githubusercontent.com/awshene/Attendance/main/confusion.png" alt="Confusion Matrix"/>
</p>

- Java
- Python
    - matplotlib
    - numpy
    - os
    - pandas
    - sklearn

- Aidan Shene
- Jake Aldridge
- Dr. Hosam Alamleh